+++ framework/master/master_boot_test.go
@@ -1,4 +1,4 @@
-package framework
+package master
 
 import (
 	"testing"
